requirements.txt
certifi==2024.2.2
charset-normalizer==2.0.12
idna==3.7
packaging==21.3
Pillow==8.4.0
pyparsing==3.0.7
python-dotenv==0.20.0
requests==2.27.1
selenium==3.141.0
urllib3==1.26.18
webdriver-manager==3.7.1
